Move Reactive Producer Interfaces to core module
